Refactor all of singleCartItems and allCartItems into cart.js redux store
